Complete Milestone 5
Added Switch row code (needs to be refined, so the node doesn't go to the end of the queue)
Added some new abilities
Updated the heal / MP cost calculations to be bound by maximums.
Added logic to healing abilities to check if the target is dead or not first. Seperated out a raise spell with high MP cost.